reorder next escalation policy, move to components
